requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-twilio
python-dotenv
gunicorn